No storage/exception handling
The scraper is pulling the overall metascore of the movie, the year that it was released in and the metascore of each individual review, the author of the review and the publication that the review was published in. I am super new to git so I goofed up and didn't update the same file. Shouldn't happen from after the next upload :P